Added lots of numeric conversion warning flags to GCC compile lines, and then went through and removed 99% of the warnings. (htons gives bogus warnings on OS X so I couldn't remove them.)
Modified the protocol to support bandwidth limits for UDP of greater than
4.3gbps. I did this by adding an 'exponent' byte into the test request
message - if the server supports the later protocol version, then this
value indicates the number of bits the existing bandwidth word should
be shifted to get the 'real' value requested. (As part of this, added
some better error checking to see if limits that are larger than can
be handled will be detected and reported as an error instead of being
munged by overflow.

Fixed what must have been a long-standing bug in the policy code. Due to
a mistake, child limit definitions have been able to specify larger limits
for things like bandwidth than the parent specified. This bug has now
been fixed - if a parent limit is more restrictive it will be used in
place of less restrictive limits specified in child limit definitions.



git-svn-id: https://svn.internet2.edu/svn/bwctl/trunk@460 2da7661c-602c-0410-9277-d9994c0c2ef1

+++ bwlib/error.c
@@ -42,7 +42,7 @@ BWLError_(
     va_start(ap,fmt);
 
     if(ctx && ctx->eh){
-        I2ErrLogVT(ctx->eh,(int)severity,etype,fmt,ap);
+        I2ErrLogVT(ctx->eh,(int)severity,(int)etype,fmt,ap);
     }
     else{
         char		buff[_BWL_ERR_MAXSTRING];
